+++ src/DTO/Messages/AnswerCallbackQueryData.php
@@ -1,0 +1,35 @@
+<?php
+
+namespace AlexItDev91\LaravelTelegramBot\DTO\Messages;
+
+use AlexItDev91\LaravelTelegramBot\DTO\Concerns\BuildsTelegramBotPayload;
+use AlexItDev91\LaravelTelegramBot\DTO\TelegramBotRequestData;
+
+final readonly class AnswerCallbackQueryData extends TelegramBotRequestData
+{
+    use BuildsTelegramBotPayload;
+
+    /**
+     * @param  array<string, mixed>  $extra
+     */
+    public function __construct(
+        string $callbackQueryId,
+        ?string $text = null,
+        ?bool $showAlert = null,
+        ?string $url = null,
+        ?int $cacheTime = null,
+        array $extra = [],
+    ) {
+        if ($cacheTime !== null) {
+            self::assertNonNegativeInteger('cache_time', $cacheTime);
+        }
+
+        parent::__construct(self::payload([
+            'callback_query_id' => $callbackQueryId,
+            'text' => $text,
+            'show_alert' => $showAlert,
+            'url' => $url,
+            'cache_time' => $cacheTime,
+        ], $extra, ['callback_query_id']));
+    }
+}